Move the methods from PredefinedShelfUtils to the service
