+++ src/main/java/com/karankumar/bookproject/backend/service/PredefinedShelfService.java
@@ -19,6 +19,7 @@ package com.karankumar.bookproject.backend.service;
 
 import com.karankumar.bookproject.backend.entity.Book;
 import com.karankumar.bookproject.backend.entity.PredefinedShelf;
+import com.karankumar.bookproject.backend.entity.Shelf;
 import com.karankumar.bookproject.backend.entity.account.User;
 import com.karankumar.bookproject.backend.repository.AuthorRepository;
 import com.karankumar.bookproject.backend.repository.BookRepository;
@@ -29,11 +30,15 @@ import lombok.extern.java.Log;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
+import java.util.HashSet;
 import java.util.List;
-import java.util.logging.Level;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static com.karankumar.bookproject.backend.util.PredefinedShelfUtils.getBooksInPredefinedShelves;
+import static com.karankumar.bookproject.backend.util.PredefinedShelfUtils.getPredefinedShelfName;
+import static com.karankumar.bookproject.backend.util.ShelfUtils.isAllBooksShelf;
 import static com.karankumar.bookproject.backend.util.TestData.generateAuthors;
 import static com.karankumar.bookproject.backend.util.TestData.generateBooks;
 import static com.karankumar.bookproject.backend.util.TestData.generateListOfTags;
@@ -150,4 +155,33 @@ public class PredefinedShelfService {
                 .collect(Collectors.toList());
     }
 
+    public List<String> getPredefinedShelfNamesAsStrings() {
+        return findAllForLoggedInUser().stream()
+                .map(Shelf::getShelfName)
+                .collect(Collectors.toList());
+    }
+
+	public Set<Book> getBooksInAllPredefinedShelves() {
+		return getBooksInPredefinedShelves(findAllForLoggedInUser());
+	}
+
+	/**
+	 * Fetches all of the books in the chosen predefined shelf
+	 */
+	public Set<Book> getBooksInChosenPredefinedShelf(String chosenShelf) {
+		Set<Book> books;
+		if (isAllBooksShelf(chosenShelf)) {
+			return getBooksInAllPredefinedShelves();
+		}
+
+		PredefinedShelf.ShelfName predefinedShelfName = getPredefinedShelfName(chosenShelf);
+		PredefinedShelf predefinedShelf =
+				findByPredefinedShelfNameAndLoggedInUser(predefinedShelfName);
+		if (predefinedShelf == null) {
+			books = new HashSet<>();
+		} else {
+			books = predefinedShelf.getBooks();
+		}
+		return books;
+	}
 }
