+++ src/main/java/com/karankumar/bookproject/ui/shelf/BooksInShelfView.java
@@ -23,7 +23,6 @@ import com.karankumar.bookproject.backend.entity.PredefinedShelf;
 import com.karankumar.bookproject.backend.service.BookService;
 import com.karankumar.bookproject.backend.service.CustomShelfService;
 import com.karankumar.bookproject.backend.service.PredefinedShelfService;
-import com.karankumar.bookproject.backend.util.PredefinedShelfUtils;
 import com.karankumar.bookproject.ui.MainView;
 import com.karankumar.bookproject.ui.book.form.BookForm;
 import com.karankumar.bookproject.ui.shelf.component.filter.AuthorFilterText;
@@ -78,15 +77,12 @@ public class BooksInShelfView extends VerticalLayout {
     private String chosenShelf;
     private final BookFilters bookFilters;
 
-    private final PredefinedShelfUtils predefinedShelfUtils;
-
     public BooksInShelfView(BookService bookService, PredefinedShelfService predefinedShelfService,
                             CustomShelfService customShelfService) {
         this.predefinedShelfService = predefinedShelfService;
-        predefinedShelfUtils = new PredefinedShelfUtils(predefinedShelfService);
         this.customShelfService = customShelfService;
         this.visibilityStrategies = initVisibilityStrategies();
-        this.bookGrid = new BookGrid(predefinedShelfUtils, customShelfService);
+        this.bookGrid = new BookGrid(customShelfService, predefinedShelfService);
         this.bookFilters = new BookFilters();
 
         this.whichShelf = new BookShelfComboBox(customShelfService);
