+++ src/test/java/com/karankumar/bookproject/backend/service/PredefinedShelfServiceTest.java
@@ -35,6 +35,8 @@ import javax.transaction.Transactional;
 import java.util.List;
 
 import java.util.Arrays;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @IntegrationTest
 @DisplayName("PredefinedShelfService should")
@@ -100,5 +102,19 @@ class PredefinedShelfServiceTest {
             softly.assertThat(predefinedShelfService.count()).isEqualTo(initialCount + 1);
             softly.assertThat(predefinedShelfService.findAllForLoggedInUser()).containsAll(expected);
         });
+    }
+
+    @Test
+    void getPredefinedShelfNamesCorrectlyAsStrings() {
+        List<String> actualShelfNames = predefinedShelfService.getPredefinedShelfNamesAsStrings();
+        List<String> expectedShelfNames =
+                Stream.of(PredefinedShelf.ShelfName.values()).map(Enum::toString).collect(
+                        Collectors.toList());
+
+        assertSoftly(softly -> {
+            softly.assertThat(actualShelfNames).hasSize(expectedShelfNames.size());
+            softly.assertThat(expectedShelfNames).containsAll(actualShelfNames);
+        });
+
     }
 }
